pkg/connutil: move CloseNotifyConn from router to connutil pkg
