+++ pkg/connutil/close_notify.go
@@ -1,9 +1,12 @@
-package proxy
+package connutil
 
 import (
 	"io"
 	"net"
+	"net/http"
 )
+
+type CloseNotifier http.CloseNotifier
 
 type closeNotifyConn struct {
 	net.Conn
