+++ router/tcp.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/flynn/flynn/Godeps/_workspace/src/golang.org/x/net/context"
 	"github.com/flynn/flynn/discoverd/cache"
+	"github.com/flynn/flynn/pkg/connutil"
 	"github.com/flynn/flynn/router/proxy"
 	"github.com/flynn/flynn/router/types"
 )
@@ -329,5 +330,5 @@ type tcpService struct {
 }
 
 func (s *tcpService) ServeConn(conn net.Conn) {
-	s.rp.ServeConn(context.Background(), proxy.CloseNotifyConn(conn))
+	s.rp.ServeConn(context.Background(), connutil.CloseNotifyConn(conn))
 }
